+++ Sources/Public/CalendarView.swift
@@ -1259,6 +1259,13 @@ extension CalendarView {
     let accessibilityScrollText = targetMonthView.accessibilityLabel
     UIAccessibility.post(notification: .pageScrolled, argument: accessibilityScrollText)
 
+    // ensure that scrolling related callbacks are still fired when performing scrolling via accessibility
+    if let visibleDayRange {
+      didScroll?(visibleDayRange, false)
+      didEndDragging?(visibleDayRange, true)
+      didEndDecelerating?(visibleDayRange)
+    }
+
     return true
   }